requirements.txt
jinja2>=3.1
google-generativeai>=0.7
gitpython>=3.1
requests>=2.31
streamlit>=1.32
python-dotenv>=1.0
pytest>=8.0
